Swagger를 통한 게시판 API 문서화

+++ example.spring/src/main/java/com/example/demo/mvc/controller/BoardController.java
@@ -3,15 +3,24 @@ package com.example.demo.mvc.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.mvc.domain.Board;
 import com.example.demo.mvc.repository.BoardRepository;
 import com.example.demo.mvc.service.BoardService;
+import com.example.demo.parameter.BoardParameter;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
 
 
 /*
@@ -20,6 +29,7 @@ import com.example.demo.mvc.service.BoardService;
  */
 @RestController
 @RequestMapping("/board")
+@Api(tags = "게시판 API")
 public class BoardController {
 	
 	@Autowired
@@ -30,6 +40,7 @@ public class BoardController {
 	 * @return
 	 */
 	@GetMapping
+	@ApiOperation(value = "목록 조회", notes = "게시물 목록 정보를 조회합니다.")
 	public List<Board> getList(){
 		return boardService.getList();
 	}
@@ -40,6 +51,10 @@ public class BoardController {
 	 * @return
 	 */
 	@GetMapping("/{boardSeq}")
+	@ApiOperation(value = "상세 조회", notes = "게시물 번호에 해당하는 상세 정보를 조회할 수 있습니다.")
+	@ApiImplicitParams({
+		@ApiImplicitParam(name = "boardSeq", value = "게시물 번호", example = "1")
+	})
 	public Board get(@PathVariable int boardSeq) {
 		return boardService.get(boardSeq);
 	}
@@ -48,19 +63,34 @@ public class BoardController {
 	 * 등록/수정 처리
 	 * @param board
 	 */
-	@GetMapping("/save")
-	public int save(Board board) {
-		boardService.save(board);
-		return boardService.save(board);
+	@PutMapping("/save")
+	@ApiOperation(value = "등록/수정 처리", notes = "신규 게시물 저장 및 기존 게시물 수정.")
+	@ApiImplicitParams({
+		@ApiImplicitParam(name = "boardSeq", value = "게시물 번호", example = "1"),
+		@ApiImplicitParam(name = "title", value = "게시물 제목", example = "changon_title"),
+		@ApiImplicitParam(name = "contents", value = "게시물 내용", example = "changon_contents"),
+	})
+	public int save(BoardParameter parameter) {
+		boardService.save(parameter);
+		return parameter.getBoardSeq();
 	}
 	
 	/**
 	 * 삭제 처리
 	 * @param boardSeq
 	 */
-	@GetMapping("/delete/{boardSeq}")
-	public void delete(@PathVariable int boardSeq) {
+	@DeleteMapping("/{boardSeq}")
+	@ApiOperation(value = "삭제 처리", notes = "게시물 번호에 해당하는 정보를 삭제.")
+	@ApiImplicitParams({
+		@ApiImplicitParam(name = "boardSeq", value = "게시물 번호", example = "1"),
+	})
+	public boolean delete(@PathVariable int boardSeq) {
+		Board board = boardService.get(boardSeq);
+		if(board == null) {
+			return false;
+		}
 		boardService.delete(boardSeq);
+		return true;
 	}
